+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Commit 1
 - Add WizardComponent sample code to the Main View.
 - Dynamically render WizardSteps based on an array in TS.
+
+# Add echo server to the application
+- It supports a POST request for /echo and return the same response as received in the request's body.
+- This allows us to control the api response from the UI itself.